requirements.txt
langchain
langchain-openai
python-dotenv
langchain_community

